requirements.txt
tensorflow>=2.15.0
pillow>=10.0.0
matplotlib>=3.7.1
scikit-learn>=1.3.0
numpy>=1.24.3